requirements.txt
flask
openai
PyPDF2
python-dotenv
selenium
webdriver-manager